Create a function to display new tweets
I cleaned up my code a bit and organized my code into functions.
I created a function that publishes the tweets to the screen, and
other two functions that utilizes the displayTweet function to
handle old tweets in the queue and new tweets.

However, this code does not diplay the tweets in reverse chronological
order, which it needs to.