Enable ability to click on user to see user's timeline
I used prepend instead of append to ensure the tweets were posted in
reverse chronological order.  Then I added a class to each div element
to be able to identify the tweet by the user's name.  This way, I am
able to filter for only the user's tweets when their username is clicked.

I was unable to use a for loop to generate on.('click') events for each
user and instead had to write each one out manually.  I'm not sure if
it's possible to generate the on.('click') events that way since it
seems to stop listening for the event after the for loop finishes.

I attempted wrapping the on.('click') events in a function to call
on later, but that didn't seem to be effective either.