Implement a click on title feature in order to return to all tweets
I added a global boolean variable (allTweets) that starts off as true,
and is set to false when we only want to see a specific user's timeline.
This variable allows us to toggle to avoid conflicts between our
on('click') functions.

I also tried using show() and hide() functions instead of removing and
re-adding, but it doesn't seem to work well after I have clicked to
one timeline, clicked back to the main stream, and then try to click
to a user's timeline again.  The tweets will start flashing on the
screen.  So it seems to work best when we add and remove instead.